connection deduplication part 1
includes some improvements to general changelog logic as well, because of course it does.

the deduplication mechanism is based on generating "clusters" from nearby events. qrect comes in very handy here, i swear c++ with qt is a whole new language at this point.

still needs some safeguards on the other end: right now the teleport destination is ignored during the clustering process. of course, connections on the opposite end will be clustered in much the same way; connections to separate worlds or parts of worlds will form separate clusters.